Correct behaviour -      int written = write(fd, buf, len); +      ssize_t written = write(fd, buf, len);
svn path=/trunk/kdelibs/; revision=144699

+++ kioslave/file/file.cc
@@ -193,7 +193,7 @@ write_all(int fd, const char *buf, size_t len)
 {
    while (len > 0)
    {
-      int written = write(fd, buf, len);
+      ssize_t written = write(fd, buf, len);
       if (written < 0)
       {
           if (errno == EINTR)
